requirements.txt
django >=1.7
django-owned-models >=0.1.1
django-shopify-webhook >=0.2.3
ShopifyAPI >=2.1.0
jsonfield >=0.9.22
setuptools >=5.7
